HW 4 (attempt 2/????)
  When doing the tutorial, what did you get stuck on?
In reading through the deck / tutorials / googling, what's not making sense? It makes sense but I’m having trouble with formatting
Same, but what DOES make sense? Is this a superior approach to coding in the 'global scope' ala code pen, or does this scoping make it more complicated? Superior
If it is more complicated, what makes it harder?
If it is a superior approach, why did you feel this was easier than the code pen based global way? I think it's superior because it gives me explanations as to what exactly I'm typing. In terms of encouragement, the code pen is nicer to my ego but it won’t have as much of an impact in the long run.